Add files via upload
#DC Measurement at Ambient Temperature
#DC Measurement After Heating the Load

In both measurements, a change in the duty cycle (D.C.) is observed due to the temperature variation of the load. Heating the load increases the duty cycle. Additionally, there is a slight difference in the duty cycle between the waveforms due to the non-ideal behavior of the operational amplifiers and transistors.